Extract html representation from cube module
